Fix the main error in the pysh lexer with case statements
I had presumed the problem to be in the parser, but that was not the case.
The lexer handles 'in' specially, since it treats it as reserved only in
specific places ('for <something> in <something else>' and 'case <something>
in'), but not reserved elsewhere.  The logic considers the third word in a for
statement as reserved if it's "in", but does not handle the case statement.
Modify to give 'case' the same treatment, since 'in' is the third word there
as well.

Signed-off-by: Chris Larson <clarson@kergoth.com>