Drop the python code unindenting, for performance reasons
The downside to this method of doing things is that it now requires that a
PythonSnippet's code be indented (they are when used from bitbake), but a
PythonValue's code does not have to be (nor should it be).  I think this is
reasonable for now, given PythonSnippet generally defines functions, but the
values are oneliners from ${@}.

I'd really prefer to keep the current implementation, but we can always return
to it if we find a better solution to the slowness of the tokenize/untokenize.

Signed-off-by: Chris Larson <chris_larson@mentor.com>